Offer BUNDLE_JAR as a synthetic compilation level
This builds each entry in your dependencies as its own BUNDLE,
and creates a single loader js to grab them all in the required
order. This avoid running a whole-world BUNDLE, allows the build
to cache much more work and parallelize each BUNDLE, but it does
cause the amount of JS loaded to be far larger. This is now the
default for test and watch, though the other options still remain.

Incremented version so that this is a distinct release, tested
separately from other changes

Temporarily added dependencyMode to allow another way to opt-out
of this change. This is already deprecated to indicate it will be
removed soon. Entrypoints will also be removed soon with this.

Also adds an integration test for elemental2 jars, to ensure that
closure has the externs that they expect and that dependencies
make sense.