Fix charset issue for test runner
When running junit.html there is no info about the content encoding of
the html file. On os with no UTF-8 standard encoding (like Windows) this
results in interpreting the file with a non-UTF-8-Encoding. This is also
true for all referenced js-files which ends up in charset issues. This
commit fix this issue by setting charset to UTF-8.