Treat pom scope=runtime as compile to make gradle poms usable
Gradle's "implementation" configuration for dependencies results in
published maven poms that are technically accurate for building
bytecode, but useless for building J2CL. To fix this, we interpret any
scope=runtime as if it says scope=compile.

This patch also tests many publicly released gwt modules to ensure that
they build cleanly in the plugin, with the hope of avoiding this issue
in the future. There are several commented out dependencies, to avoid
trying to build gwt-dom until the next release is ready. Bug #127 is
filed for followup.

Fixes #126