Treat pom scope=runtime as compile to make gradle poms usable
Gradle's "implementation" configuration for dependencies results in
published maven poms that are technically accurate for building
bytecode, but useless for building J2CL. To fix this, we interpret any
scope=runtime as if it says scope=compile.

A separate commit will include a test case to verify that the published
gwt modules in maven central work properly.

Fixes #126