Update to released jsinterop-base, remove all other repos (#144)
As of this commit there are no snapshot dependencies, we should be
nearly ready for a release.

Due to those snapshot repositories (our own, mostly), there was a race
condition that somehow we got lucky with before, and circleci happened
to get the plugin ready for us before we needed it. The CI process now
always runs install, so that the archetype integration tests can
access the newly-built plugin.