Update to latest closure v20220502-1
This is the end of persistent input store, which was never used in this
iteration of j2cl-m-p. Instead, we'll be moving towards the typedast
feature, which closure-compiler uses to break its build into "stages".

This also fixes #157, by adding a new commit to our closure-compiler
feature of allowing sourcemaps in BUNDLE mode - contents loaded from a
different fs root will not try to relativize any longer.