Add optional ability to use Turbine  (#165)
Turbine is a header compiler for Java that allows us to reduce the
source code to parse, leading to decreased compile time. By ignoring
changes in the body, we can also avoid downstream builds that can be
skipped.

This feature is optional for now, but may be set to the default in a
future release. To enable this, assign the taskMappings key of
stripped_bytecode_headers to the value turbine.

Co-authored-by: Colin Alworth <colin@vertispan.com>
Fixes #150