Changed my mind. 10s staleness, and only check external processes
This means that a jvm shouldn't be able to race itself for staleness,
but will only race other JVMs, including ones that have already exited.
There is a risk here of one JVM being delayed, or the FS not correctly
reflected "last modified time" between two JVMs - we might need a more
approriate channel with which to communicate.