Rough draft of bundling sourcemaps and original sources
This will cost extra CPU time, but save on IO time and disk space, which
is likely a net win for most builds. This is a first draft, which
appears to work as expected (sourcemaps function correctly, vastly
smaller output in BUNDLE_JAR), but doesn't yet fully reduce the number
of copies we can make, and doesn't avoid copying sourcemap bundles (in
cases where js bundles are not copied). There may also be caching
issues, where sourcemaps are incorrect not generated or are regenerated.